requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.74
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# Data acquisition and processing
yfinance>=0.2.28
finnhub-python>=2.4.20
tavily-python>=0.5.0
stockstats>=0.6.2

# Vector database and memory
chromadb>=0.5.0
openai>=1.50.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
requests>=2.31.0

# Visualization and monitoring
rich>=13.0.0
matplotlib>=3.7.0
plotly>=5.17.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0

# Type hints and validation
pydantic>=2.5.0
typing-extensions>=4.8.0

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Utilities
tenacity>=8.2.0
psutil>=5.9.0
